Small modification to rarefaction function.
It appears it was trying to convert a 2d vector into a list and it was vastly inefficient. Converting and returning it as a numeric vector allowed the compile to optimize out a lot of noise and make it much faster. Conversions take time, therefore it may be smart to ensure we are returning objects as R objects and not C objects.